2023 11 6

+++ trading/AutoTraderOld/AutoOrderDefaultHandler.java
@@ -1,5 +1,6 @@
 package AutoTraderOld;
 
+import client.Decimal;
 import client.OrderState;
 import client.OrderStatus;
 import controller.ApiController;
@@ -121,7 +122,8 @@ public class AutoOrderDefaultHandler implements ApiController.IOrderHandler {
     }
 
     @Override
-    public void orderStatus(OrderStatus status, double filled, double remaining, double avgFillPrice, int permId, int parentId, double lastFillPrice, int clientId, String whyHeld, double mktCapPrice) {
+    public void orderStatus(OrderStatus status, Decimal filled, Decimal remaining, double avgFillPrice,
+                            int permId, int parentId, double lastFillPrice, int clientId, String whyHeld, double mktCapPrice) {
         outputDetailedXU(ibContractToSymbol(AutoTraderXU.activeFutCt)
                 , str("||OrderStatus||", defaultID,
                         globalIdOrderMap.get(defaultID), status, filled,
