+++ trading/api/OrderAugmented.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import static DevTrader.BreachTrader.f2;
+import static Trader.BreachTrader.f2;
 import static utility.Utility.ibContractToSymbol;
 import static utility.Utility.str;
 
@@ -63,7 +63,7 @@ public class OrderAugmented {
         }
 
         double tradedPrice = order.lmtPrice();
-        double size = order.totalQuantity();
+        double size = order.totalQuantity().longValue();
         if (order.action() == Types.Action.BUY) {
             return Math.round(100d * (currPrice - tradedPrice - costPerUnit) * size) / 100d;
         } else if (order.action() == Types.Action.SELL) {
