+++ trading/handler/SGXPositionHandler.java
@@ -1,6 +1,7 @@
 package handler;
 
 import client.Contract;
+import client.Decimal;
 import controller.ApiController;
 import historical.HistChinaStocks;
 
@@ -8,10 +9,10 @@ import javax.swing.*;
 
 public class SGXPositionHandler implements ApiController.IPositionHandler {
         @Override
-        public void position(String account, Contract contract, double position, double avgCost) {
+        public void position(String account, Contract contract, Decimal position, double avgCost) {
             String ticker = utility.Utility.ibContractToSymbol(contract);
             SwingUtilities.invokeLater(() -> {
-                HistChinaStocks.currentPositionMap.put(ticker, (int) position);
+                HistChinaStocks.currentPositionMap.put(ticker, position);
             });
         }
